pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "incidentlens-env"
version = "0.1.0"
description = "IncidentLens: Log analysis and incident triage environment for OpenEnv"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "fastapi>=0.100.0",
    "openai>=2.0.0",
    "openenv-core[core]>=0.2.3",
    "pydantic>=2.0.0",
    "uvicorn[standard]>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.24.0",
]

[project.scripts]
server = "incidentlens_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = [
    "incidentlens_env",
    "incidentlens_env.server",
    "incidentlens_env.tasks",
]
package-dir = { "incidentlens_env" = ".", "incidentlens_env.server" = "server", "incidentlens_env.tasks" = "tasks" }
